# swagger.yaml
openapi: 3.0.0
info:
  title: Gadgets API
  version: 1.0.0
  description: API for managing gadgets with authentication

servers:
  - url: https://imfgadgetapi.onrender.com
    description: Production server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
          format: password
      required:
        - username
        - password

    Gadget:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [Available, Deployed, Decommissioned, Destroyed]

paths:
  /auth/register:
    post:
      summary: Register a new user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User registered successfully
        '400':
          description: Username and password are required
        '500':
          description: Internal Server Error

  /auth/login:
    post:
      summary: Login user
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '401':
          description: Invalid credentials
        '500':
          description: Internal Server Error

  /gadgets:
    post:
      summary: Create a new gadget
      tags:
        - Gadgets
      security:
        - BearerAuth: []
      responses:
        '201':
          description: Gadget created successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error

    get:
      summary: Get all gadgets
      tags:
        - Gadgets
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
          description: Filter gadgets by status
      responses:
        '200':
          description: List of gadgets retrieved successfully
        '401':
          description: Unauthorized
        '500':
          description: Internal Server Error

  /gadgets/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: Gadget ID

    patch:
      summary: Update gadget status
      tags:
        - Gadgets
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [Available, Deployed]
      responses:
        '200':
          description: Gadget status updated successfully
        '400':
          description: Invalid status
        '401':
          description: Unauthorized
        '404':
          description: Gadget not found
        '500':
          description: Internal Server Error

    delete:
      summary: Decommission a gadget
      tags:
        - Gadgets
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Gadget decommissioned successfully
        '401':
          description: Unauthorized
        '404':
          description: Gadget not found
        '500':
          description: Internal Server Error

  /gadgets/{id}/self-destruct:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
        description: Gadget ID

    post:
      summary: Self-destruct a gadget
      tags:
        - Gadgets
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                confirmationCode:
                  type: string
                  description: 6-digit confirmation code
      responses:
        '200':
          description: Self-destruct sequence completed successfully
        '400':
          description: Invalid confirmation code
        '401':
          description: Unauthorized
        '404':
          description: Gadget not found
        '500':
          description: Internal Server Error